better ai
it follows the ball every bounce (after an arbitrary "reaction speed" delay). this is instead of every frame

was mostly good practice to code, but does feels a bit better to play i think. paddles moves smoothly to where it thinks the ball will go, minus bounces. ideally this teaches player to do the same thing, and to beat it, predict bounces